Moved error routes into separate module
